[MM-61123] Add diagnostics data to the Support Packet (#36)

+++ server/app/statistics.go
@@ -3,6 +3,6 @@
 
 package app
 
-func (a *App) GetUsedCardsCount() (int, error) {
+func (a *App) GetUsedCardsCount() (int64, error) {
 	return a.store.GetUsedCardsCount()
 }
